python/requirements.txt
# Usage: pip install -r requirements.txt
ultralytics>=8.2.34
flask>=3.0.3
pika>=1.3.2
minio>=7.2.8
dill>=0.3.8

tqdm==4.66.5
opencv-python==4.6.0.66
pyyaml==6.0.2
torch==2.4.1
numpy==1.24.4
pillow==10.4.0
requests==2.32.3
psutil==6.0.0
matplotlib==3.7.5
scipy==1.10.1
werkzeug==3.0.4
schedule==1.2.2
psutil==6.0.0
GPUtil==1.4.0

mysql-connector-python~=9.0.0
PyYAML~=6.0.1
shapely~=2.0.7
pandas~=2.0.3
albumentations~=1.4.18